requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose
passlib[bcrypt]
pydantic
